requirements.txt
py-cord==2.6.1
python-dotenv>=1.0.0
colorlog==6.9.0
psutil>=5.9.7
tqdm==4.67.1
telethon>=1.33.1
cryptg==0.4.0  # For faster encryption/decryption
pyaes==1.6.1  # Required by Telethon
rsa==4.9.1  # Required by Telethon

# Monitoring
prometheus-client==0.19.0

# Security
pyjwt==2.8.0
cryptography==42.0.5
passlib==1.7.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pytest-mock==3.12.0
pytest-timeout>=2.2.0  # For more reliable tests with timeouts
pytest>=8.0.0

# Development
black>=23.11.0
mypy>=1.7.1
flake8>=6.1.0
isort==5.13.2
aiofiles>=23.2.1

# API and Data
openai>=1.3.7
redis>=5.0.1
pyyaml>=6.0.1

# Visualization and reporting
matplotlib>=3.8.0  # For log visualization and reports
fpdf>=1.7.2  # For PDF generation
numpy>=1.26.0  # Required by matplotlib

# Utilities
aiohttp>=3.9.1  # Async HTTP client
zoneinfo>=0.2.1;python_version<"3.9"  # For timezone handling (built-in for Python 3.9+) 